+++ src/lib/server/locks/handle-messages.ts
@@ -16,7 +16,7 @@ import { building } from '$app/environment';
 import { paymentMethods } from '../payment-methods';
 import { userQuery } from '../user';
 import { rateLimit } from '../rateLimit';
-import { computePriceInfo } from '$lib/types/Cart';
+import { computePriceInfo, freeProductUnitsForCart } from '$lib/cart';
 import { filterNullish } from '$lib/utils/fillterNullish';
 import type { Price } from '$lib/types/Order';
 import { sendAuthentificationlink } from '../sendNotification';
@@ -552,13 +552,14 @@ const commands: Record<
 							: [];
 
 						const totalPrice = computePriceInfo(items, {
-							deliveryFees: { amount: 0, currency: 'SAT' },
-							vatExempted: runtimeConfig.vatExempted,
-							userCountry: undefined,
 							bebopCountry: runtimeConfig.vatCountry || undefined,
+							deliveryFees: { amount: 0, currency: 'SAT' },
+							freeProductUnits: freeProductUnitsForCart(cart.items),
+							userCountry: undefined,
+							vatExempted: runtimeConfig.vatExempted,
 							vatNullOutsideSellerCountry: runtimeConfig.vatNullOutsideSellerCountry,
-							vatSingleCountry: runtimeConfig.vatSingleCountry,
-							vatProfiles
+							vatProfiles,
+							vatSingleCountry: runtimeConfig.vatSingleCountry
 						});
 
 						return paymentMethods({
